Step 3, Lab 2: correction XXI… + Setting of the locale was moved to correct early place! + Translation! Everything dialog was not in Swedish! + Integer to float in roulettegame implicitly!

+++ step3/1dv434.s3.l2/roulettegame.cpp
@@ -25,7 +25,7 @@ int RouletteGame::play(char* a_bet, int a_amount) {
 		}
 		++m_shootCount;
 
-		return a_amount * m_shootCount * 1.5f;
+		return (int) (a_amount * m_shootCount * 1.5f);
 	}
 
 	return a_amount;
